requirements.txt
flask>=2.3
python-dotenv>=1.0
openai>=1.0
gunicorn>=21.0